backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
python-dotenv
